fix(imu): address CodeRabbit review findings (round 13)
- Fix BMP280 IIR filter register value: 0x08->0x04 (was coefficient 4, not 2 as documented)
- Rename proto ImuData.calib_stat->calibration_status (field 15) for clarity; regenerate nanopb
- Update telemetry_task.c to use new calibration_status field name
- Add second _Static_assert in main.c enforcing BNO055/BMP280 I2C address distinctness
- Add @pre/@post to main() in test_rx_bmp280.c
- Fix test coverage table in test_rx_bno055.c (Init=6, Read=4, Calibration=3)
- Add IMU loop overrun warning log in imu_task.c when elapsed >= period
- Add explicit (ULONG) casts to tx_thread_sleep calls in rx_bno055.c

+++ star-rx72n-firmware/libs/rx_bno055/src/rx_bno055.c
@@ -453,8 +453,8 @@ static rx_err_t internal_init_reset_and_wait(void)
     rx_log_error(s_tag, "Software reset failed");
     return err;
   }
-  (void)tx_thread_sleep(k_bno055_delay_por_ms /
-                        k_bno055_ms_per_tick); /* 65 ticks @ 10 ms/tick = 650 ms */
+  (void)tx_thread_sleep(
+    (ULONG)(k_bno055_delay_por_ms / k_bno055_ms_per_tick)); /* 65 ticks @ 10 ms/tick = 650 ms */
 
   /* Step 2: Enter CONFIG mode (required for configuration register writes) */
   err = internal_write_reg((uint8_t)k_bno055_reg_opr_mode, (uint8_t)k_bno055_opr_config);
@@ -463,8 +463,8 @@ static rx_err_t internal_init_reset_and_wait(void)
     return err;
   }
   (void)tx_thread_sleep(
-    k_bno055_delay_config_ms / k_bno055_ms_per_tick +
-    k_bno055_tick_round_up); /* 2 ticks @ 10 ms/tick = 20 ms (round up for 19 ms) */
+    (ULONG)(k_bno055_delay_config_ms / k_bno055_ms_per_tick +
+            k_bno055_tick_round_up)); /* 2 ticks @ 10 ms/tick = 20 ms (round up for 19 ms) */
 
   return k_rx_ok;
 }
